+++ packages/retry-plugin/src/types.d.ts
@@ -1,5 +1,5 @@
 export interface FetchWithRetryOptions {
-  url?: string;
+  url: string;
   options?: RequestInit;
   retryTimes?: number;
   retryDelay?: number;
@@ -7,22 +7,13 @@ export interface FetchWithRetryOptions {
 }
 
 export interface ScriptWithRetryOptions {
-  url?: string;
-  attrs?: Record<string, string>;
   retryTimes?: number;
   retryDelay?: number;
-  customCreateScript?: CreateScriptFunc;
+  moduleName?: Array<string>;
+  cb?: (resolve: (value: unknown) => void, error: any) => void;
 }
 
 export type RetryPluginParams = {
-  fetch?: FetchWithRetryOptions; // fetch retry options
-  script?: ScriptWithRetryOptions; // script retry options
+  fetch?: FetchWithRetryOptions;
+  script?: ScriptWithRetryOptions;
 };
-
-export type CreateScriptFunc = (
-  url: string,
-  attrs: Record<string, any>,
-) => HTMLScriptElement;
-
-export type RequiredUrl<T extends { url?: string }> = Required<Pick<T, 'url'>> &
-  Omit<T, 'url'>;