+++ packages/chrome-devtools/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @module-federation/devtools
 
+## 0.6.14
+
+### Patch Changes
+
+- Updated dependencies [ad605d2]
+  - @module-federation/sdk@0.6.14
+
 ## 0.6.13
 
 ### Patch Changes