+++ packages/data-prefetch/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @module-federation/data-prefetch
 
+## 0.6.14
+
+### Patch Changes
+
+- ad605d2: chore: unified logger
+- Updated dependencies [ad605d2]
+  - @module-federation/runtime@0.6.14
+  - @module-federation/sdk@0.6.14
+
 ## 0.6.13
 
 ### Patch Changes