+++ packages/enhanced/CHANGELOG.md
@@ -1,5 +1,21 @@
 # [0.2.0-canary.5](https://github.com/module-federation/core/compare/enhanced-0.2.0-canary.4...enhanced-0.2.0-canary.5) (2023-11-20)
 
+## 0.6.14
+
+### Patch Changes
+
+- ad605d2: chore: unified logger
+- Updated dependencies [ad605d2]
+- Updated dependencies [86b4cdc]
+  - @module-federation/data-prefetch@0.6.14
+  - @module-federation/dts-plugin@0.6.14
+  - @module-federation/managers@0.6.14
+  - @module-federation/manifest@0.6.14
+  - @module-federation/sdk@0.6.14
+  - @module-federation/rspack@0.6.14
+  - @module-federation/runtime-tools@0.6.14
+  - @module-federation/bridge-react-webpack-plugin@0.6.14
+
 ## 0.6.13
 
 ### Patch Changes